Flytter aktivitetsplikt ut i egen route.
Forberedelse til nye endepunkter.

EY-3820
